django-api/requirements.txt
Django==3.2.13
djangorestframework==3.13.1
PyJWT==2.4.0
django-cors-headers==3.13.0
gunicorn==20.1.0
